pyproject.toml
[project]
name = "cubepython"
version = "0.1.0"
description = "MTG Cube recommender engine with backend API and data scraping"
requires-python = ">=3.8"
dependencies = [
    # Backend dependencies
    "fastapi==0.115.5",
    "uvicorn[standard]==0.32.1",
    "sqlmodel==0.0.22",
    "aiosqlite==0.20.0",
    "pydantic==2.10.3",
    "pydantic-settings==2.6.1",
    "httpx==0.28.1",
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    # Scripts dependencies
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "selenium>=4.15.0",
    "python-dotenv>=1.0.0",
    "boto3>=1.28.0",
]
